+++ lib/bt/stage_specification.rb
@@ -1,4 +1,4 @@
-require 'bt/psych'
+require 'bt/yaml'
 
 module BT
   class GeneratedSpecification < Struct.new :file
